# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  types.LoginReq:
    properties:
      address:
        type: string
      chain_id:
        type: integer
      message:
        type: string
      signature:
        type: string
    type: object
  types.UserLoginResp:
    properties:
      result: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@nftauction.com
    name: API Support
    url: http://www.nftauction.com/support
  description: NFT Auction Backend API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NFT Auction API
  version: "1.0"
paths:
  /collections/{address}/{token_id}:
    get:
      consumes:
      - application/json
      description: 获取指定物品的详细信息
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 物品Token ID
        in: path
        name: token_id
        required: true
        type: string
      - description: 链ID
        in: query
        name: chain_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 物品详情
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取物品详情
      tags:
      - collections
  /collections/{address}/{token_id}/bids:
    get:
      consumes:
      - application/json
      description: 获取集合中指定物品的所有出价信息
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 物品Token ID
        in: path
        name: token_id
        required: true
        type: string
      - description: 过滤参数，JSON格式
        in: query
        name: filters
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 物品出价信息列表
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取集合中特定物品的出价信息
      tags:
      - collections
  /collections/{address}/{token_id}/traits:
    get:
      consumes:
      - application/json
      description: 获取指定物品的特性(Attribute)信息
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 物品Token ID
        in: path
        name: token_id
        required: true
        type: string
      - description: 链ID
        in: query
        name: chain_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 物品特性信息
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取物品特性信息
      tags:
      - collections
  /collections/{address}/bids:
    get:
      consumes:
      - application/json
      description: 获取指定集合的所有出价信息
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 过滤参数，JSON格式
        in: query
        name: filters
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 出价信息列表
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取集合的出价信息
      tags:
      - collections
  /collections/{address}/history-sales:
    get:
      consumes:
      - application/json
      description: 获取指定集合的历史销售价格信息
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 链ID
        in: query
        name: chain_id
        required: true
        type: integer
      - default: 7d
        description: 时间范围(24h/7d/30d)
        in: query
        name: duration
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 历史销售价格信息
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取历史销售价格信息
      tags:
      - collections
  /collections/{address}/items:
    get:
      consumes:
      - application/json
      description: 根据过滤条件获取指定集合中的NFT物品列表
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 过滤参数，JSON格式
        in: query
        name: filters
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 物品列表
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取集合中的物品列表
      tags:
      - collections
  /collections/{address}/top-trait:
    get:
      consumes:
      - application/json
      description: 获取指定集合中物品特性的最高价格信息
      parameters:
      - description: 集合地址
        in: path
        name: address
        required: true
        type: string
      - description: 过滤参数，JSON格式
        in: query
        name: filters
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 特性价格信息
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取物品特性的最高价格信息
      tags:
      - collections
  /collections/createNft:
    post:
      consumes:
      - application/json
      description: 铸造nft
      parameters:
      - description: nft名称
        in: query
        name: name
        required: true
        type: string
      - description: NFT描述
        in: query
        name: description
        required: true
        type: string
      - description: NFT图片URL
        in: query
        name: imageUrl
        required: true
        type: string
      - description: 版税百分比
        in: query
        name: royaltyPercentage
        required: true
        type: string
      - description: 所属区块链网络id
        in: query
        name: chainId
        required: true
        type: string
      - description: 类别id
        in: query
        name: categorieId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 出价信息列表
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 铸造nft
      tags:
      - collections
  /portfolio/bid-orders:
    post:
      consumes:
      - application/json
      description: 批量查询出价信息
      parameters:
      - description: 过滤参数，JSON格式
        in: query
        name: filters
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 出价信息列表
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 批量查询出价信息
      tags:
      - user
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 接收用户上传的附件（图片、视频等）并保存到本地
      parameters:
      - description: 上传的文件（支持JPG/PNG/GIF/MP4/MP3）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '成功返回示例：{''status'': ''success'', ''path'': ''/uploads/123.jpg''}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '错误示例：{''error'': ''File too large''}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '错误示例：{''error'': ''Failed to save file''}'
          schema:
            additionalProperties: true
            type: object
      summary: 上传文件（NFT附件）
      tags:
      - NFT
  /user/{address}/login-message:
    get:
      consumes:
      - application/json
      description: 获取用户登录所需的签名信息
      parameters:
      - description: 用户地址
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录签名信息
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取登录签名信息
      tags:
      - user
  /user/{address}/sig-status:
    get:
      consumes:
      - application/json
      description: 获取指定用户的签名状态信息
      parameters:
      - description: 用户地址
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 签名状态信息
          schema: {}
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取用户签名状态
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户通过签名信息登录系统
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功响应
          schema:
            $ref: '#/definitions/types.UserLoginResp'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 用户登录qaq
      tags:
      - user
schemes:
- http
- https
swagger: "2.0"